requirements.txt
selenium
beautifulsoup4
python-telegram-bot
webdriver-manager
requests
python-dotenv
undetected-chromedriver
setuptools
